+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the AyeCode Template Manager will be documented in this f
 
 ## [3.0.3-beta] - 2026-04-TBD
 - Fatal error if loaded twice - FIXED
+- Implemented AyeCode Package Loader - ADDED
 
 ## [3.0.2-beta] - 2026-04-09
 - composer package type wrong - FIXED